# feat: find_entry simd optimization