--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,6 +1,9 @@
 // Copyright 2021-2022 Parity Technologies (UK) Ltd.
 // This file is dual-licensed as Apache-2.0 or MIT.
 
+#![cfg_attr(nightly, feature(portable_simd))]
+#![cfg_attr(nightly, feature(test))]
+
 mod btree;
 mod column;
 mod compress;