This PR tries to address https://github.com/paritytech/parity-db/issues/171 by using [std::simd](https://doc.rust-lang.org/nightly/std/simd/) on nightly rust.

My local stress test results:
`RUSTFLAGS="-Ctarget-feature=+avx2" cargo +nightly run -p parity-db-admin --release -- stress`
```
sse2
Completed 100000 commits in 70.011662619 seconds. 1428.3334555871788 cps. 0 hits, 0 misses, 0 iterations, 0 qps
Completed 10000000 queries in 6.168205102 seconds. 1621217.1668477051 qps

base
Completed 100000 commits in 62.011042791 seconds. 1612.616003524352 cps. 0 hits, 0 misses, 0 iterations, 0 qps
Completed 10000000 queries in 5.826384161 seconds. 1716330.3557868504 qps

base_simd 2 lanes
Completed 100000 commits in 66.00952211 seconds. 1514.9329491184221 cps. 0 hits, 0 misses, 0 iterations, 0 qps
Completed 10000000 queries in 5.915917492 seconds. 1690354.8796146733 qps

base_simd 4 lanes
Completed 100000 commits in 63.01187526 seconds. 1587.0024433867327 cps. 0 hits, 0 misses, 0 iterations, 0 qps
Completed 10000000 queries in 5.814712547 seconds. 1719775.4694098036 qps
```