@arkpar 2023-02-06
Thank you, but I'm afraid we can't accept it. This library is mainly used in the stable rustc environment. Also, maintenance costs and extra dependencies are just not worth the marginal difference in query performance. 
I've also added some benchmarks in https://github.com/paritytech/parity-db/pull/180. Even though the function itself is twice as fast, the query performance (qps) is only 3-5% better when measured in the stress test. 

> sse2
Completed 100000 commits in 70.011662619 seconds. 1428.3334555871788 cps. 0 hits, 0 misses, 0 iterations, 0 qps
Completed 10000000 queries in 6.168205102 seconds. 1621217.1668477051 qps

> base
Completed 100000 commits in 62.011042791 seconds. 1612.616003524352 cps. 0 hits, 0 misses, 0 iterations, 0 qps
Completed 10000000 queries in 5.826384161 seconds. 1716330.3557868504 qps

On my machine sse2 was a bit faster than base when averaged over 5 runs each.